Removed duplicated code, removed rsciio_velox specific concept mapping code to make this a shared function now for all parsers in the em parser who need it, reorganized location of configuration files to convey cleaner how the components of this parser package that em is works and tries to avoid redundancy where possible
