+++ pynxtools_em/config/README.md
@@ -1,0 +1,9 @@
+# Context
+
+Configuration how concepts defined in specific serialization formats are mapped on concepts in NeXus if possible:
+* eln_cfg.py - NOMAD Oasis ELN export file format
+* oasis_cfg.py - Specific configuration of the em parser in a specific local NOMAD Oasis
+  to avoid having to enter these pieces of information with every ELN entry again
+* image_png_protochips_cfg.py - PNG files exported with AXONStudio software from Protochips tech partner
+* image_tiff_tfs_cfg.py - TIFF files exported with a ThermoFisher Apreo SEM from ThermoFisher/FEI tech partner
+* rsciio_velox_cfg.py - Velox EMD file format from ThermoFisher/FEI tech partner
