+++ pynxtools_em/config/image_png_protochips_cfg.py
