+++ pynxtools_em/examples/README.md
@@ -1,4 +1,8 @@
 # Context
 
-Specific code that is relevant only for the implementation of examples for
-this em parser and the NOMAD OASIS research data management system.
+Specific code that is relevant only for the implementation of specific examples
+contributed by members of the community to implement specific work within
+the NOMAD Oasis research data management system.
+
+Users of code in this directory should consider that possibly modifications are
+required to make the em parser fit for the specific tasks they wish to perform.
