+++ pynxtools_em/geometry/msmse_convention.py
@@ -20,8 +20,6 @@
 suggested by D. Rowenhorst et al. in DOI: 10.1088/0965-0393/23/8/083501.
 """
 
-# pylint: disable=no-member,duplicate-code
-
 msmse_convention = {
     "three_dimensional_rotation_handedness": "counter_clockwise",
     "rotation_convention": "passive",
