+++ pynxtools_em/utils/get_gitrepo_commit.py
@@ -1,0 +1,38 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""Get (last) commit of the git repository."""
+
+import os
+from subprocess import CalledProcessError, run
+
+
+def get_repo_last_commit() -> str:
+    """Identify the last commit to the repository."""
+    try:
+        return (
+            run(
+                ["git", "describe", "--always"],
+                cwd=os.getcwd(),
+                check=True,
+                capture_output=True,
+            )
+            .stdout.decode("utf-8")
+            .strip()
+        )
+    except (FileNotFoundError, CalledProcessError):
+        return None
