+++ pynxtools_em/utils/io_case_logic.py
@@ -21,9 +21,6 @@
 
 from typing import Tuple, Dict, List
 
-from pynxtools_em.shared.mapping_functors import variadic_path_to_specific_path
-from pynxtools_em.shared.shared_utils import get_sha256_of_file_content
-
 VALID_FILE_NAME_SUFFIX_CONFIG = [".yaml", ".yml"]
 VALID_FILE_NAME_SUFFIX_DATA = [".emd"]
 # ".png", ".tiff", ".dm3", ".dm4", ".hdf5", ".h5", ".dream3d", ".mtex", ".nion"]
