+++ pynxtools_em/utils/nx_default_plots.py
