+++ pynxtools_em/utils/string_encoding.py
